+++ common/src/main/java/com/iamkaf/torchtoss/ThrowableTorchItem.java
@@ -1,0 +1,9 @@
+package com.iamkaf.torchtoss;
+
+import net.minecraft.world.item.SnowballItem;
+
+public class ThrowableTorchItem extends SnowballItem {
+    public ThrowableTorchItem(Properties properties) {
+        super(properties);
+    }
+}