+++ changelog.md
@@ -1,9 +1,21 @@
-# Template Changelog
+# Torch Toss Changelog
 
-See the full changelog at https://github.com/iamkaf/template-mod/commits/main/
+See the full changelog at https://github.com/iamkaf/torch-toss/commits/main/
+
+## 1.2.0
+
+- feat: added the new torches to the functional blocks creative tab
+- feat(1.21.1): 1.21.1 port
+
+## 1.1.0
+
+- feat: added throwable soul torches
+- feat: added throwable redstone torches
+
+## 1.0.1
+
+- fix: snowballs no longer place down torches (oops)
 
 ## 1.0.0
 
 Initial Release
-
-##